isTrusted extension logic
